Refactor sync logic and optimize backend history fetching
- Extracted `syncBitunixHistory` from `src/services/app.ts` to new `src/services/syncService.ts` to improve maintainability and decouple logic.
- Increased pagination limit in `src/routes/api/sync/orders/+server.ts` from 20 to 100 pages to support deeper order history.
- Improved timestamp parsing robustness in backend order sync to prevent infinite loops on invalid data.
- Addressed security vulnerabilities via `npm audit fix`.
